(BSR)[PRO] refactor: move utils to their parent component
